+++ src/test/kotlin/dev/meanmail/parsing/ParsingTest.kt
@@ -71,6 +71,10 @@ class ParsingTest : ParsingTestCase(
         doTest(true)
     }
 
+    fun testIfEscapedParens() {
+        doTest(true)
+    }
+
     override fun getTestDataPath(): String {
         return "src/test/resources/" +
                 this.javaClass.packageName.replace(".", "/")